refactor: merge OpenCds healthcheck into ServiceChain module
