--- apps/server-nestjs/src/modules/opencds/opencds.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common'
-import { TerminusModule } from '@nestjs/terminus'
-import { ConfigurationModule } from '../infrastructure/configuration/configuration.module'
-import { OpenCdsHealthService } from './opencds-health.service'
-
-@Module({
-  imports: [ConfigurationModule, TerminusModule],
-  providers: [OpenCdsHealthService],
-  exports: [OpenCdsHealthService],
-})
-export class OpenCdsModule {}
