+++ apps/server-nestjs/src/modules/service-chain/service-chain.module.ts
@@ -5,6 +5,7 @@ import { DatabaseModule } from '../infrastructure/database/database.module'
 import { EventsModule } from '../infrastructure/events/events.module'
 import { UserPermissionModule } from '../infrastructure/permission/user/user.module'
 import { OpenCdsClientService } from './open-cds-client.service'
+import { ServiceChainHealthService } from './service-chain-health.service'
 import { ServiceChainController } from './service-chain.controller'
 import { ServiceChainService } from './service-chain.service'
 
@@ -17,7 +18,7 @@ import { ServiceChainService } from './service-chain.service'
     UserPermissionModule,
   ],
   controllers: [ServiceChainController],
-  providers: [OpenCdsClientService, ServiceChainService],
-  exports: [ServiceChainService],
+  providers: [OpenCdsClientService, ServiceChainHealthService, ServiceChainService],
+  exports: [ServiceChainService, ServiceChainHealthService],
 })
 export class ServiceChainModule {}
